Moved regions into their own folder
