+++ pol-core/pol/CMakeSources.cmake
@@ -96,8 +96,6 @@ set (pol_sources  # sorted !
   globals/ucfg.h
   globals/uvars.cpp
   globals/uvars.h
-  guardrgn.cpp
-  guardrgn.h
   guilds.cpp
   guilds.h
   guildscrobj.cpp
@@ -137,8 +135,6 @@ set (pol_sources  # sorted !
   menu.h
   menusel.cpp
   miscmsg.cpp
-  miscrgn.cpp
-  miscrgn.h
   mkscrobj.cpp
   mkscrobj.h
   mobile/attack.cpp
@@ -228,8 +224,6 @@ set (pol_sources  # sorted !
   multi/multidef.h
   multi/multidef2.cpp
   multi/multis.cpp
-  musicrgn.cpp
-  musicrgn.h
   network/auxclient.cpp
   network/auxclient.h
   network/bannedips.cpp
@@ -323,13 +317,19 @@ set (pol_sources  # sorted !
   realms/realmlos.cpp
   reftypes.cpp
   reftypes.h
-  region.cpp
-  region.h
+  regions/region.cpp
+  regions/region.h
+  regions/guardrgn.cpp
+  regions/guardrgn.h
+  regions/miscrgn.cpp
+  regions/miscrgn.h
+  regions/musicrgn.cpp
+  regions/musicrgn.h
+  regions/resource.cpp
+  regions/resource.h
   repsys.cpp
   repsys.h
   repsys_cfg.h
-  resource.cpp
-  resource.h
   savedata.cpp
   savedata.h
   schedule.cpp
