+++ pol-core/pol/globals/uvars.cpp
@@ -27,24 +27,24 @@
 #include "../accounts/accounts.h"
 #include "../checkpnt.h"
 #include "../console.h"
-#include "../guardrgn.h"
+#include "regions/guardrgn.h"
+#include "regions/musicrgn.h"
+#include "regions/resource.h"
+#include "regions/miscrgn.h"
 #include "../guilds.h"
 #include "../item/equipmnt.h"
 #include "../item/itemdesc.h"
 #include "../item/weapon.h"
 #include "../listenpt.h"
 #include "../loadunld.h"
-#include "../miscrgn.h"
 #include "../mobile/attribute.h"
 #include "../multi/boat.h"
-#include "../musicrgn.h"
 #include "../npctmpl.h"
 #include "../objecthash.h"
 #include "../party.h"
 #include "../polcfg.h"
 #include "../polsem.h"
-#include "../realms/realm.h"
-#include "../resource.h"
+#include "realms/realm.h"
 #include "../scrstore.h"
 #include "../spells.h"
 #include "../startloc.h"
