+++ pol-core/pol/mobile/charactr.cpp
@@ -113,7 +113,9 @@
 #include "../globals/settings.h"
 #include "../globals/state.h"
 #include "../globals/uvars.h"
-#include "../guardrgn.h"
+#include "regions/guardrgn.h"
+#include "regions/miscrgn.h"
+#include "regions/musicrgn.h"
 #include "../guilds.h"
 #include "../item/armor.h"
 #include "../item/item.h"
@@ -122,7 +124,6 @@
 #include "../item/wepntmpl.h"
 #include "../layers.h"
 #include "../mdelta.h"
-#include "../miscrgn.h"
 #include "../mkscrobj.h"
 #include "../module/uomod.h"
 #include "../movecost.h"
@@ -130,7 +131,6 @@
 #include "../multi/house.h"
 #include "../multi/multi.h"
 #include "../multi/multidef.h"
-#include "../musicrgn.h"
 #include "../network/cgdata.h"
 #include "../network/client.h"
 #include "../network/cliface.h"
