+++ pol-core/pol/module/uomod.cpp
@@ -104,7 +104,9 @@
 #include "../gameclck.h"
 #include "../globals/object_storage.h"
 #include "../globals/uvars.h"
-#include "../guardrgn.h"
+#include "regions/guardrgn.h"
+#include "regions/miscrgn.h"
+#include "regions/resource.h"
 #include "../item/item.h"
 #include "../item/itemdesc.h"
 #include "../layers.h"
@@ -113,7 +115,6 @@
 #include "../los.h"
 #include "../mdelta.h"
 #include "../menu.h"
-#include "../miscrgn.h"
 #include "../mobile/charactr.h"
 #include "../mobile/npc.h"
 #include "../mobile/ufacing.h"
@@ -137,7 +138,6 @@
 #include "../profile.h"
 #include "realms/realms.h"
 #include "realms/realm.h"
-#include "../resource.h"
 #include "../savedata.h"
 #include "../scrdef.h"
 #include "../scrsched.h"
