+++ pol-core/pol/regions/guardrgn.cpp
