+++ pol-core/pol/regions/musicrgn.cpp
@@ -6,8 +6,9 @@
 
 #include "musicrgn.h"
 
-#include "../clib/cfgelem.h"
-#include "../clib/random.h"
+#include "clib/cfgelem.h"
+#include "clib/random.h"
+
 #include "globals/uvars.h"
 
 namespace Pol
