+++ pol-core/pol/regions/region.cpp
@@ -11,13 +11,14 @@
 #include <string>
 #include <tuple>
 
-#include "../bscript/berror.h"
-#include "../bscript/impstr.h"
-#include "../clib/cfgelem.h"
-#include "../clib/cfgfile.h"
-#include "../clib/fileutil.h"
-#include "../clib/stlutil.h"
-#include "../plib/poltype.h"
+#include "bscript/berror.h"
+#include "bscript/impstr.h"
+#include "clib/cfgelem.h"
+#include "clib/cfgfile.h"
+#include "clib/fileutil.h"
+#include "clib/stlutil.h"
+#include "plib/poltype.h"
+
 #include "globals/uvars.h"
 #include "proplist.h"
 #include "realms/realms.h"
