+++ pol-core/pol/regions/resource.cpp
