+++ pol-core/pol/uimport.cpp
@@ -54,7 +54,7 @@
 #include "multi/multi.h"
 #include "objecthash.h"
 #include "polvar.h"
-#include "resource.h"
+#include "regions/resource.h"
 #include "savedata.h"
 #include "servdesc.h"
 #include "startloc.h"
